+++ src/exojax/spec/layeropacity.py
@@ -1,0 +1,113 @@
+"""compute opacity difference in atmospheric layers
+
+"""
+
+from jax import jit, vmap
+import jax.numpy as jnp
+from exojax.spec.hitrancia import interp_logacia_matrix
+from exojax.spec.hminus import log_hminus_continuum
+from exojax.atm.idealgas import number_density
+from exojax.utils.constants import logkB, logm_ucgs
+from exojax.utils.constants import opfac
+from exojax.spec.dtau_mmwl import dtauM_mmwl
+
+
+def layer_optical_depth(dParr, xsm, MR, mass, g):
+    """dtau of the molecular cross section.
+
+    Note:
+       opfac=bar_cgs/(m_u (g)). m_u: atomic mass unit. It can be obtained by fac=1.e3/m_u, where m_u = scipy.constants.m_u.
+
+    Args:
+       dParr: delta pressure profile (bar) [N_layer]
+       xsm: cross section matrix (cm2) [N_layer, N_nus]
+       MR: volume mixing ratio (VMR) or mass mixing ratio (MMR) [N_layer]
+       mass: mean molecular weight for VMR or molecular mass for MMR
+       g: gravity (cm/s2)
+
+    Returns:
+       optical depth matrix [N_layer, N_nus]
+    """
+
+    return opfac * xsm * dParr[:, None] * MR[:, None] / (mass * g)
+
+
+def layer_optical_depth_CIA(nus, Tarr, Parr, dParr, vmr1, vmr2, mmw, g, nucia, tcia, logac):
+    """dtau of the CIA continuum. Not used in art.
+
+    Args:
+       nus: wavenumber matrix (cm-1)
+       Tarr: temperature array (K)
+       Parr: temperature array (bar)
+       dParr: delta temperature array (bar)
+       vmr1: volume mixing ratio (VMR) for molecules 1 [N_layer]
+       vmr2: volume mixing ratio (VMR) for molecules 2 [N_layer]
+       mmw: mean molecular weight of atmosphere
+       g: gravity (cm2/s)
+       nucia: wavenumber array for CIA
+       tcia: temperature array for CIA
+       logac: log10(absorption coefficient of CIA)
+
+    Returns:
+       optical depth matrix  [N_layer, N_nus]
+    """
+    narr = number_density(Parr, Tarr)
+    lognarr1 = jnp.log10(vmr1 * narr)  # log number density
+    lognarr2 = jnp.log10(vmr2 * narr)  # log number density
+    logg = jnp.log10(g)
+    ddParr = dParr / Parr
+    dtauc = (10**(interp_logacia_matrix(Tarr, nus, nucia, tcia, logac) +
+                  lognarr1[:, None] + lognarr2[:, None] + logkB - logg -
+                  logm_ucgs) * Tarr[:, None] / mmw * ddParr[:, None])
+
+    return dtauc
+
+
+def layer_optical_depth_VALD(dParr, xsm, VMR, mmw, g):
+    """dtau of the atomic (+ionic) cross section from VALD.
+
+    Args:
+       dParr: delta pressure profile (bar) [N_layer]
+       xsm: cross section matrix (cm2) [N_species x N_layer x N_wav]
+       VMR: volume mixing ratio [N_species x N_layer]
+       mmw: mean molecular weight [N_layer]
+       g: gravity (cm/s2)
+
+    Returns:
+        dtau: optical depth matrix [N_layer x N_nus]
+    """
+    dtauS = jit(vmap(dtauM_mmwl, (None, 0, 0, None, None)))( \
+                            dParr, xsm, VMR, mmw, g)
+    dtau = jnp.abs(jnp.sum(dtauS, axis=0))
+    return dtau
+
+
+def layer_optical_depth_Hminus(nus, Tarr, Parr, dParr, vmre, vmrh, mmw, g):
+    """dtau of the H- continuum.
+
+    Args:
+       nus: wavenumber matrix (cm-1)
+       Tarr: temperature array (K)
+       Parr: temperature array (bar)
+       dParr: delta temperature array (bar)
+       vmre: volume mixing ratio (VMR) for e- [N_layer]
+       vmrH: volume mixing ratio (VMR) for H atoms [N_layer]
+       mmw: mean molecular weight of atmosphere
+       g: gravity (cm2/s)
+
+    Returns:
+       optical depth matrix  [N_layer, N_nus]
+    """
+    narr = number_density(Parr, Tarr)
+    #       number_density_e: number density for e- [N_layer]
+    #       number_density_h: number density for H atoms [N_layer]
+    number_density_e = vmre * narr
+    number_density_h = vmrh * narr
+    logg = jnp.log10(g)
+    ddParr = dParr / Parr
+    logabc = (log_hminus_continuum(nus, Tarr, number_density_e,
+                                   number_density_h))
+    dtauh = 10**(logabc + logkB - logg -
+                 logm_ucgs) * Tarr[:, None] / mmw * ddParr[:, None]
+
+    return dtauh
