+++ src/exojax/spec/rtransfer.py
@@ -1,22 +1,32 @@
 """Radiative transfer module used in exospectral analysis."""
 from jax import jit, vmap
 import jax.numpy as jnp
-import numpy as np
 from exojax.special.expn import E1
-from exojax.spec.hitrancia import interp_logacia_matrix
-from exojax.spec.hminus import log_hminus_continuum
-from exojax.atm.idealgas import number_density
-from exojax.utils.constants import logkB, logm_ucgs
-from exojax.utils.constants import opfac
 import warnings
+from exojax.spec.layeropacity import layer_optical_depth
+from exojax.spec.layeropacity import layer_optical_depth_CIA 
+from exojax.spec.layeropacity import layer_optical_depth_Hminus 
+from exojax.spec.layeropacity import layer_optical_depth_VALD 
 
+def dtauM(dParr, xsm, MR, mass, g):
+   warn_msg = "Use `spec.layeropacity.layer_optical_depth` instead"
+   warnings.warn(warn_msg, FutureWarning)
+   return layer_optical_depth(dParr, xsm, MR, mass, g)
 
-def wavenumber_grid(x0, x1, N, unit='cm-1', xsmode='lpf'):
-    warn_msg = "Use `utils.grids.wavenumber_grid` instead"
-    warnings.warn(warn_msg, FutureWarning)
-    from exojax.utils.grids import wavenumber_grid
-    return wavenumber_grid(x0, x1, N, unit=unit, xsmode=xsmode)
+def dtauCIA(nus, Tarr, Parr, dParr, vmr1, vmr2, mmw, g, nucia, tcia, logac): 
+   warn_msg = "Use `spec.layeropacity.layer_optical_depth_CIA` instead"
+   warnings.warn(warn_msg, FutureWarning)
+   return layer_optical_depth_CIA(nus, Tarr, Parr, dParr, vmr1, vmr2, mmw, g, nucia, tcia, logac)
 
+def dtauHminus(nus, Tarr, Parr, dParr, vmre, vmrh, mmw, g):
+   warn_msg = "Use `spec.layeropacity.layer_optical_depth_Hminus` instead"
+   warnings.warn(warn_msg, FutureWarning)
+   return layer_optical_depth_Hminus(nus, Tarr, Parr, dParr, vmre, vmrh, mmw, g)
+
+def dtauVALD(dParr, xsm, VMR, mmw, g):
+   warn_msg = "Use `spec.layeropacity.layer_optical_depth_VALD` instead"
+   warnings.warn(warn_msg, FutureWarning)
+   return layer_optical_depth_VALD(dParr, xsm, VMR, mmw, g)
 
 def pressure_layer(log_pressure_top=-8.,
                    log_pressure_btm=2.,
@@ -29,193 +39,6 @@ def pressure_layer(log_pressure_top=-8.,
     return pressure_layer_logspace(log_pressure_top, log_pressure_btm, NP,
                                    mode, numpy)
 
-
-def dtauCIA(nus, Tarr, Parr, dParr, vmr1, vmr2, mmw, g, nucia, tcia, logac):
-    """dtau of the CIA continuum.
-
-    Args:
-       nus: wavenumber matrix (cm-1)
-       Tarr: temperature array (K)
-       Parr: temperature array (bar)
-       dParr: delta temperature array (bar)
-       vmr1: volume mixing ratio (VMR) for molecules 1 [N_layer]
-       vmr2: volume mixing ratio (VMR) for molecules 2 [N_layer]
-       mmw: mean molecular weight of atmosphere
-       g: gravity (cm2/s)
-       nucia: wavenumber array for CIA
-       tcia: temperature array for CIA
-       logac: log10(absorption coefficient of CIA)
-
-    Returns:
-       optical depth matrix  [N_layer, N_nus]
-    """
-    narr = number_density(Parr, Tarr)
-    lognarr1 = jnp.log10(vmr1 * narr)  # log number density
-    lognarr2 = jnp.log10(vmr2 * narr)  # log number density
-    logg = jnp.log10(g)
-    ddParr = dParr / Parr
-    dtauc = (10**(interp_logacia_matrix(Tarr, nus, nucia, tcia, logac) +
-                  lognarr1[:, None] + lognarr2[:, None] + logkB - logg -
-                  logm_ucgs) * Tarr[:, None] / mmw * ddParr[:, None])
-
-    return dtauc
-
-
-def dtauCIA_mmwl(nus, Tarr, Parr, dParr, vmr1, vmr2, mmw, g, nucia, tcia,
-                 logac):
-    """dtau of the CIA continuum.
-       (for the case where mmw is given for each atmospheric layer)
-
-    Args:
-       nus: wavenumber matrix (cm-1)
-       Tarr: temperature array (K)
-       Parr: temperature array (bar)
-       dParr: delta temperature array (bar)
-       vmr1: volume mixing ratio (VMR) for molecules 1 [N_layer]
-       vmr2: volume mixing ratio (VMR) for molecules 2 [N_layer]
-       mmw: mean molecular weight of atmosphere [N_layer]
-       g: gravity (cm2/s)
-       nucia: wavenumber array for CIA
-       tcia: temperature array for CIA
-       logac: log10(absorption coefficient of CIA)
-
-    Returns:
-       optical depth matrix  [N_layer, N_nus]
-    """
-    narr = number_density(Parr, Tarr)
-    lognarr1 = jnp.log10(vmr1 * narr)  # log number density
-    lognarr2 = jnp.log10(vmr2 * narr)  # log number density
-    logg = jnp.log10(g)
-    ddParr = dParr / Parr
-    dtauc = (10**(interp_logacia_matrix(Tarr, nus, nucia, tcia, logac) +
-                  lognarr1[:, None] + lognarr2[:, None] + logkB - logg -
-                  logm_ucgs) * Tarr[:, None] / mmw[:, None] * ddParr[:, None])
-
-    return dtauc
-
-
-def dtauM(dParr, xsm, MR, mass, g):
-    """dtau of the molecular cross section.
-
-    Note:
-       opfac=bar_cgs/(m_u (g)). m_u: atomic mass unit. It can be obtained by fac=1.e3/m_u, where m_u = scipy.constants.m_u.
-
-    Args:
-       dParr: delta pressure profile (bar) [N_layer]
-       xsm: cross section matrix (cm2) [N_layer, N_nus]
-       MR: volume mixing ratio (VMR) or mass mixing ratio (MMR) [N_layer]
-       mass: mean molecular weight for VMR or molecular mass for MMR
-       g: gravity (cm/s2)
-
-    Returns:
-       optical depth matrix [N_layer, N_nus]
-    """
-
-    return opfac * xsm * dParr[:, None] * MR[:, None] / (mass * g)
-
-
-def dtauM_mmwl(dParr, xsm, MR, mass, g):
-    """dtau of the molecular cross section.
-       (for the case where mmw is given for each atmospheric layer)
-
-    Note:
-       opfac=bar_cgs/(m_u (g)). m_u: atomic mass unit. It can be obtained by fac=1.e3/m_u, where m_u = scipy.constants.m_u.
-
-    Args:
-       dParr: delta pressure profile (bar) [N_layer]
-       xsm: cross section matrix (cm2) [N_layer, N_nus]
-       MR: volume mixing ratio (VMR) or mass mixing ratio (MMR) [N_layer]
-       mass: mean molecular weight for VMR or molecular mass for MMR [N_layer]
-       g: gravity (cm/s2)
-
-    Returns:
-       optical depth matrix [N_layer, N_nus]
-    """
-
-    return opfac * xsm * dParr[:, None] * MR[:, None] / (mass[:, None] * g)
-
-
-@jit
-def dtauVALD(dParr, xsm, VMR, mmw, g):
-    """dtau of the atomic (+ionic) cross section from VALD.
-
-    Args:
-       dParr: delta pressure profile (bar) [N_layer]
-       xsm: cross section matrix (cm2) [N_species x N_layer x N_wav]
-       VMR: volume mixing ratio [N_species x N_layer]
-       mmw: mean molecular weight [N_layer]
-       g: gravity (cm/s2)
-
-    Returns:
-        dtau: optical depth matrix [N_layer x N_nus]
-    """
-    dtauS = jit(vmap(dtauM_mmwl, (None, 0, 0, None, None)))( \
-                            dParr, xsm, VMR, mmw, g)
-    dtau = jnp.abs(jnp.sum(dtauS, axis=0))
-    return dtau
-
-
-def dtauHminus(nus, Tarr, Parr, dParr, vmre, vmrh, mmw, g):
-    """dtau of the H- continuum.
-
-    Args:
-       nus: wavenumber matrix (cm-1)
-       Tarr: temperature array (K)
-       Parr: temperature array (bar)
-       dParr: delta temperature array (bar)
-       vmre: volume mixing ratio (VMR) for e- [N_layer]
-       vmrH: volume mixing ratio (VMR) for H atoms [N_layer]
-       mmw: mean molecular weight of atmosphere
-       g: gravity (cm2/s)
-
-    Returns:
-       optical depth matrix  [N_layer, N_nus]
-    """
-    narr = number_density(Parr, Tarr)
-    #       number_density_e: number density for e- [N_layer]
-    #       number_density_h: number density for H atoms [N_layer]
-    number_density_e = vmre * narr
-    number_density_h = vmrh * narr
-    logg = jnp.log10(g)
-    ddParr = dParr / Parr
-    logabc = (log_hminus_continuum(nus, Tarr, number_density_e,
-                                   number_density_h))
-    dtauh = 10**(logabc + logkB - logg -
-                 logm_ucgs) * Tarr[:, None] / mmw * ddParr[:, None]
-
-    return dtauh
-
-
-def dtauHminus_mmwl(nus, Tarr, Parr, dParr, vmre, vmrh, mmw, g):
-    """dtau of the H- continuum.
-       (for the case where mmw is given for each atmospheric layer)
-
-    Args:
-       nus: wavenumber matrix (cm-1)
-       Tarr: temperature array (K)
-       Parr: temperature array (bar)
-       dParr: delta temperature array (bar)
-       vmre: volume mixing ratio (VMR) for e- [N_layer]
-       vmrH: volume mixing ratio (VMR) for H atoms [N_layer]
-       mmw: mean molecular weight of atmosphere [N_layer]
-       g: gravity (cm2/s)
-
-    Returns:
-       optical depth matrix  [N_layer, N_nus]
-    """
-    narr = number_density(Parr, Tarr)
-    #       number_density_e: number density for e- [N_layer]
-    #       number_density_h: number density for H atoms [N_layer]
-    number_density_e = vmre * narr
-    number_density_h = vmrh * narr
-    logg = jnp.log10(g)
-    ddParr = dParr / Parr
-    logabc = (log_hminus_continuum(nus, Tarr, number_density_e,
-                                   number_density_h))
-    dtauh = 10**(logabc + logkB - logg -
-                 logm_ucgs) * Tarr[:, None] / mmw[:, None] * ddParr[:, None]
-
-    return dtauh
 
 
 @jit
@@ -235,8 +58,12 @@ def trans2E3(x):
     return (1.0 - x) * jnp.exp(-x) + x**2 * E1(x)
 
 
-@jit
 def rtrun(dtau, S):
+   warnings.warn("Use rtrun_emis_pure_absorption instead", FutureWarning)
+   return rtrun_emis_pure_absorption(dtau, S)
+
+@jit
+def rtrun_emis_pure_absorption(dtau, S):
     """Radiative Transfer using two-stream approximaion + 2E3 (Helios-R1 type)
 
     Args:
@@ -253,9 +80,8 @@ def rtrun(dtau, S):
                    jnp.cumprod(jnp.vstack([jnp.ones(Nnus), TransM]), axis=0),
                    axis=0)
 
-
 @jit
-def rtrun_surface(dtau, S, Sb):
+def rtrun_emis_pure_absorption_surface(dtau, S, Sb):
     """Radiative Transfer using two-stream approximaion + 2E3 (Helios-R1 type)
     with a planetary surface.
 
@@ -276,7 +102,7 @@ def rtrun_surface(dtau, S, Sb):
 
 
 @jit
-def rtrun_direct(dtau, S):
+def rtrun_emis_pure_absorption_direct(dtau, S):
     """Radiative Transfer using direct integration.
 
     Note:
@@ -291,3 +117,5 @@ def rtrun_direct(dtau, S):
     """
     taupmu = jnp.cumsum(dtau, axis=0)
     return jnp.sum(S * jnp.exp(-taupmu) * dtau, axis=0)
+
+def rtrun_trans_pure_absorption(dtau_chord, S):
