+++ tests/integration/modit_scanfft/modit_scanfft_test.py
@@ -52,7 +52,7 @@ def test_rt_exomol():
     from exojax.spec.modit import exomol
     from exojax.spec.modit_scanfft import xsmatrix_scanfft
     from exojax.spec.rtransfer import dtauM
-    from exojax.spec.rtransfer import rtrun
+    from exojax.spec.rtransfer import rtrun_emis_pure_absorption
     from exojax.spec.planck import piBarr
     from exojax.spec.modit import set_ditgrid_matrix_exomol
     from exojax.test.data import TESTDATA_CO_EXOMOL_MODIT_EMISSION_REF
@@ -80,7 +80,7 @@ def test_rt_exomol():
                    nus, dgm_ngammaL)
     dtau = dtauM(dParr, jnp.abs(xsm), MMR * np.ones_like(Parr), molmass, g)
     sourcef = piBarr(Tarr, nus)
-    F0 = rtrun(dtau, sourcef)
+    F0 = rtrun_emis_pure_absorption(dtau, sourcef)
     filename = pkg_resources.resource_filename(
         'exojax', 'data/testdata/' + TESTDATA_CO_EXOMOL_MODIT_EMISSION_REF)
     dat = pd.read_csv(filename, delimiter=",", names=("nus", "flux"))
@@ -140,7 +140,7 @@ def test_rt_vald():
     dtauatom = rt.dtauVALD(dParr, xsmS, VMR_uspecies, mmw*ONEARR, g)
 
     sourcef = piBarr(Tarr, nus)
-    F0 = rt.rtrun(dtauatom, sourcef)
+    F0 = rt.rtrun_emis_pure_absorption(dtauatom, sourcef)
     # The reference data was generated by
     #
     # >>> np.savetxt("modit_rt_test_vald_ref.txt",np.array([nus,F0]).T,delimiter=",")
