+++ tests/unittests/spec/atmrt/atmrt_test.py
@@ -1,6 +1,6 @@
-import pytest
 from exojax.spec.atmrt import ArtCommon
 from exojax.spec.atmrt import ArtEmisPure
+from exojax.spec.atmrt import ArtTransPure
 from exojax.test.emulate_mdb import mock_wavenumber_grid
 
 def test_ArtCommon():
@@ -10,6 +10,10 @@ def test_ArtCommon():
 def test_ArtEmisPure():
     nu_grid, wav, res = mock_wavenumber_grid()
     art = ArtEmisPure(nu_grid,pressure_top=1.e-8, pressure_btm=1.e2, nlayer=100)
+
+def test_ArtTransPure()
+    nu_grid, wav, res = mock_wavenumber_grid()
+    art = ArtTransPure(nu_grid,pressure_top=1.e-8, pressure_btm=1.e2, nlayer=100)
 
 if __name__ == "__main__":
     test_ArtCommon()
