+++ tests/unittests/spec/transmission/opachord_test.py
@@ -2,31 +2,8 @@ import pytest
 import jax.numpy as jnp
 import numpy as np
 from exojax.atm.atmprof import normalized_layer_height
-from jax import jit
-#normalized_radius = normalized_radius_profile(temperature, pressure, dParr,
-#                                              mmw, radius_btm, gravity_btm)
-
-@jit
-def chord_geometric_matrix(height, radius):
-    """compute chord geometric matrix
-
-    Args:
-        height (1D array): (normalized) height of the layers from top atmosphere, Nlayer
-        radius (1D array): (normalized) radius of the layers from top atmosphere, Nlayer
-
-    Returns:
-        2D array: chord geometric matrix (Nlayer, Nlayer)
-    """
-    radius_roll = jnp.roll(radius, 1)
-
-    # elements at the top layer to be zero
-    radius_roll = radius_roll.at[0].set(radius[0])
-    height = height.at[0].set(jnp.inf)
-
-    fac_right = jnp.sqrt(radius[None, :]**2 - radius[:, None]**2)
-    fac_left = jnp.sqrt(radius_roll[None, :]**2 - radius[:, None]**2)
-    raw_matrix = (fac_left - fac_right) / height
-    return jnp.tril(raw_matrix)
+from exojax.spec.opachord import chord_geometric_matrix
+from exojax.spec.opachord import tauchord
 
 
 def test_chord_geometric_matrix():
@@ -34,33 +11,7 @@ def test_chord_geometric_matrix():
     height = 0.1 * jnp.ones(Nlayer)
     radius = jnp.cumsum(height)[::-1] + 1.0
     cgm = chord_geometric_matrix(height, radius)
-    assert jnp.sum(cgm) == pytest.approx(32.06652)
-
-
-def tauchord(chord_geometric_matrix, xsmatrix):
-    """chord opacity vector from a chord geometric matrix and xsmatrix
-    
-    Args:
-        chord_geometric_matrix (jnp array): chord geometric matrix (Nlayer, Nlayer), lower triangle matrix 
-        xsmatrix (jnp array): cross section matrix (Nlayer, N_wavenumber)
-
-    Returns: tauchord matrix (Nlayer, N_wavenumber)
-
-    """
-    return jnp.dot(chord_geometric_matrix, xsmatrix)
-
-
-def inverse_cumsum(arr):
-    return jnp.cumsum(arr[::-1])[::-1]
-
-
-def test_inverse_cumsum():
-    Tarr = jnp.array([0, 1, 2, 3, 4])
-
-    vec = inverse_cumsum(Tarr)
-
-    ans = jnp.array([10, 10, 9, 7, 4])
-    assert jnp.all(vec == ans)
+    assert jnp.sum(cgm) == pytest.approx(32.06652 * 2.0)
 
 
 def test_check_parallel_Ax_tauchord():
@@ -99,7 +50,6 @@ def test_first_layer_height_from_compute_normalized_radius_profile():
 
 
 if __name__ == "__main__":
-    #test_check_parallel_Ax_tauchord()
-    #test_inverse_cumsum()
-    #test_first_layer_height_from_compute_normalized_radius_profile()
+    test_check_parallel_Ax_tauchord()
+    test_first_layer_height_from_compute_normalized_radius_profile()
     test_chord_geometric_matrix()
